refactor(engine): split reeds.ts into focused reeds/ modules
Splits the 564-line comment service into 5 focused modules:
- reeds/types.ts (CommentRecord, CommentSettingsRecord, etc.)
- reeds/queries.ts (reads, threading, settings lookup)
- reeds/mutations.ts (create, edit, delete, moderate, settings upsert)
- reeds/blocked.ts (user blocking — raw D1, cross-tenant)
- reeds/rate-limiting.ts (comment rate limits — raw D1, cross-tenant)

Original reeds.ts becomes a thin re-export barrel — all existing
import paths continue to work unchanged.
