+++ libs/engine/src/lib/server/services/reeds/index.ts
@@ -1,0 +1,38 @@
+export type {
+	CommentRecord,
+	CommentSettingsRecord,
+	ThreadedComment,
+	BlockedCommenterRecord,
+} from "./types.js";
+
+export {
+	getCommentSettings,
+	getApprovedComments,
+	getPrivateReplies,
+	getPendingComments,
+	getPendingCount,
+	getModeratedComments,
+	getAllPrivateReplies,
+	getCommentCount,
+	getCommentById,
+	buildCommentTree,
+} from "./queries.js";
+
+export {
+	stripControlChars,
+	createComment,
+	isWithinEditWindow,
+	editComment,
+	deleteComment,
+	moderateComment,
+	upsertCommentSettings,
+} from "./mutations.js";
+
+export {
+	isUserBlocked,
+	blockCommenter,
+	unblockCommenter,
+	getBlockedCommenters,
+} from "./blocked.js";
+
+export { checkCommentRateLimit } from "./rate-limiting.js";
